[Agent] Split placeholder resolver utilities
